pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keyveve-takehome"
version = "0.1.0"
description = "Agentic Travel Planner - Keyveve Take-Home"
authors = [{name = "Engineering Team", email = "eng@keyveve.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "types-python-dateutil",
    "hypothesis>=6.88.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.0",
    # PR2: Database infrastructure
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "redis>=5.0.0",
    # PR4: LangGraph orchestrator + SSE + UI
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",
    "streamlit>=1.39.0",
    "sse-starlette>=2.1.0",
    # Chat interface
    "openai>=1.0.0",
    # PR13: MCP Integration
    "httpx>=0.27.0",  # Already included, but ensuring MCP client support
    # PR14: Authentication & Security
    "argon2-cffi>=23.1.0",      # Password hashing
    "pyjwt[crypto]>=2.8.0",     # JWT with RSA support
    "cryptography>=41.0.0",     # RSA key handling
    "python-multipart>=0.0.6",  # Form data parsing
    "email-validator>=2.0.0",   # Email validation for Pydantic
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
]

[tool.setuptools]
packages = ["backend"]
package-dir = {"" = "."}

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=backend",
    "--cov-report=term-missing",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "unit: marks tests as unit tests (deselect with '-m \"not unit\"')",
    "integration: marks tests as integration tests",
    "eval: marks tests as evaluation tests",
]
